only checking novnc.pid for vnc based vis

+++ stop.sh
@@ -6,7 +6,6 @@ echo $id
 echo "removing docker container $id"
 docker stop $id && docker rm $id && echo "container removed"
 
-
 list_descendants ()
 {
   local children=$(ps -o pid= --ppid "$1")
@@ -19,6 +18,8 @@ list_descendants ()
   echo "$children"
 }
 
-echo "killing novnc and all children"
-pid=$(cat novnc.pid)
-kill $pid $(list_descendants $pid)
+if [ -f novnc.pid ]; then
+    echo "killing novnc and all children"
+    pid=$(cat novnc.pid)
+    kill $pid $(list_descendants $pid)
+fi
